requirements.txt
flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
ollama>=0.1.0
websockets>=11.0.0
markdown>=3.4.0
python-dateutil>=2.8.0